requirements.txt
Flask>=1.1.2
Flask-Cors>=3.0.8
Flask-Bcrypt>=0.7.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
flask-restplus>=0.13.0
Flask-SQLAlchemy>=2.4.4
gunicorn>=20.0.4
psycopg2-binary>=2.8.5
PyJWT>=1.7.1
pytest>=6.2.2
pytest-dotenv>=0.5.2
python-dotenv>=0.14.0
requests>=2.24.0
Werkzeug==0.16.1